SCALE-SEAM (83): the banner that ran to the end of the file
// --- CX-1 commissioning loop --- covered 44 methods, of which THREE
were commissioning. Nine out of client.ts (1,339 -> 1,292) into three
existing mixins; no new withX() on ApiClient, every call site unchanged.

  is commissioning closing, in detail?   cxSeed, cxMatrix, cxDossier
                                         -> documents.ts
  does this model pass its checks?       ciRun, ciLatest, rulesRun,
                                         rulesGeometryRun, rebarCheckCage
                                         -> model.ts
  what does it cost, by storey?          qtoByFloor -> estimate.ts

The cx trio belongs beside closeoutSummary, which already returns a
commissioning ROLLUP (cx_count / passed / failed / conditional) - these
three are the detail behind that number, the same pair as
procurementLevel / bidLevelingDetail at ⓽. cxSeed is a POST that creates
records and came anyway: separating a seeder from the only two readers of
what it seeds is the split (82) refused when it kept namingAudit with
namingConventions.

Two placements were decided by RETURN SHAPE, against what the route said.
qtoByFloor reads as a quantity method on /qto/by-floor, but every line
carries rate and amount and the payload has a grand_total - priced
takeoff, so it went to estimating rather than to the model reads. And
rebarCheckCage came to model.ts while its two /rebar/ siblings did NOT:
rebarBbs and rebarBbsCsvUrl share a route prefix and a REBAR-RULES doc tag
with it and answer a different question - a bar bending schedule is a
quantity, an ACI cage check is the same question envelopeAudit asks.
Route-and-tag kinship is exactly the reason ⓽ nearly filed
bidLevelingDetail in ai.ts.

The banner was REPLACED, not narrowed. With the three /cx/* methods gone
it named nothing at all - one step past the over-claiming that cost ⓽, ⓾,
(81) and (82) a method each. In its place is an UNFILED map of the 35 that
remain, grouped by what each cluster answers, including five marked as
STAYING: enumOptions, searchAll, attachmentUrl, templates and
rvtBridgeStatus are client-level, not domain. Those are the first methods
in this sequence marked as not to be extracted - a decomposition that
never says "this one is already home" will eventually move something that
should not move.

Also corrected two stale counts in docs/roadmap.md found while editing it:
the CX-1 banner was recorded as 49 methods (it is 44, counted), and the
item line still said one unnumbered slice had shipped when (81), (82) and
(83) make three.

ModelCiReport moved to model.ts's type imports and out of client.ts's.

Verification: tsc --noEmit exit 0; full web suite 202 files / 2,056 tests;
eslint clean; test_file_sizes pin lowered 1,339 -> 1,292 and exact;
roadmapLanes 17/17; test_claude_md_gates (300 citations, 3 docs);
test_changelog_current; ruff clean.
