+++ docs/roadmap.md
@@ -1188,7 +1188,7 @@ two rows share a path, so two agents in different rows cannot collide.
 | **F · Docs & demo** | `README.md`, `docs/`, `apps/web/src/demo/` | keep the shipped surface honest (below) — no coded items. **`demoData.test.ts` now gates the shell's startup endpoints**; re-run `build_demo_data.py` and that test after adding one |
 | **G · API surface** | `services/api/src/aec_api/routers/`, `main.py` | no standalone items: **every lane routes its own work**, which is why this is a lane rather than a shared file |
 | **H · Registers** | `services/api/modules/*/module.json` | — |
-| **I · API client** | `apps/web/src/api/` | SCALE-SEAM *(the only open slice; ②–⓾ have shipped. **Deliberately carries NO mark**: ⓾ is the last glyph in `MARKS` and the double-circled range it closes has no successor, so the next slice cannot be numbered at all — writing a mark the parser does not know would drop the item out of this table's own population.)* *(the cell named ⓽ until 2026-09-03; ②–⓼ had shipped. This cell named ⑬–⑳ until 2026-08-24 — eight slices whose extractions had already landed — because the item regex could not see `㉒` at all, so nothing required this row to be right)* |
+| **I · API client** | `apps/web/src/api/` | SCALE-SEAM *(the only open slice; ②–⓾ plus (81)–(83) have shipped. **Deliberately carries NO mark**: ⓾ is the last glyph in `MARKS` and the double-circled range it closes has no successor, so the next slice cannot be numbered at all — writing a mark the parser does not know would drop the item out of this table's own population.)* *(the cell named ⓽ until 2026-09-03; ②–⓼ had shipped. This cell named ⑬–⑳ until 2026-08-24 — eight slices whose extractions had already landed — because the item regex could not see `㉒` at all, so nothing required this row to be right)* |
 | **J · Build & tooling** | `apps/web/scripts/`, `apps/web/vite.config.ts`, `apps/web/src/style.css`, `apps/web/src/tooling/`, `services/api/test_file_sizes.py`, `services/api/run_tests.py` | R39-TSC-CACHE *(local typecheck once diverged from CI; cause unknown, prior explanation retracted — an OBSERVATION, not a defect with a known fix. Read the entry before "fixing" it: the proposed fix is named there and rejected)* |
 
 **Parked — not available to pick up.** These are decisions or multi-release commitments, listed so
@@ -3061,19 +3061,32 @@ verbs, with a command bar as the escape hatch to everything); and **role-shaped 
 
 ## 🧱 Decomposition & reliability carry-overs (interleave one per few releases)
 
-- ◧ ⭐ **SCALE-SEAM — `client.ts` is no longer a god-file, but the split is not finished.** *(②–⓾ plus one unnumbered slice have shipped, **⓺ onward in the v0.3.1143 follow-on** — no version bump, tag lag already at the `test_release_current` bound)*
+- ◧ ⭐ **SCALE-SEAM — `client.ts` is no longer a god-file, but the split is not finished.** *(②–⓾ plus three unnumbered slices — (81), (82), (83) — have shipped, **⓺ onward in the v0.3.1143 follow-on** — no version bump, tag lag already at the `test_release_current` bound)*
   **THE NEXT SLICE NEEDS A NEW GLYPH SCHEME.** ⓾ (U+24FE) is the last glyph in `MARKS`
   (`apps/web/src/shell/roadmapLanes.test.ts`), which runs ①–⑳, ㉑–㊿, ❶–❿, ⓫–⓴, ⓵–⓾. The
   double-circled range ⓵–⓾ is now EXHAUSTED, and the item code here
   carries no mark for the same reason — `roadmapLanes.test.ts` parses the mark against `MARKS`,
   so writing one it does not know would drop this item out of the lane population entirely.
-  **`client.ts` is 1,339 lines.** The RACI banner is dealt with — it covered 13 methods and
+  **`client.ts` is 1,292 lines.** The RACI banner is dealt with — it covered 13 methods and
   described 4; the five model-quality audits went to `model.ts`, `namingAudit` to
   `documents.ts` beside its `namingConventions` pair, and the three remaining non-RACI methods
   (`mcpTools`, `handoverAcceptance`, `inspectVim`) are now under an explicit UNFILED header
-  naming them. **The big one left is `// --- CX-1 commissioning loop ---` at 49 methods**,
-  among them `rebarBbs` and `rebarCheckCage` — roughly a quarter of the file under a banner
-  for something else.
+  naming them.
+
+  **The CX-1 banner is dealt with too, and it was the largest instance of the class.**
+  `// --- CX-1 commissioning loop ---` ran to the END of the file over **44** methods — this
+  entry said 49 until (83) counted them — of which **three** were commissioning. (83) took
+  those three to `documents.ts` (they detail the `commissioning` rollup `closeoutSummary`
+  already returns), the model-check quartet plus `rebarCheckCage` to `model.ts`, and
+  `qtoByFloor` to `estimate.ts`. With the `/cx/*` methods gone the banner named **nothing**, so
+  it was replaced by an UNFILED map of the 35 that remain rather than narrowed.
+
+  **What is left is that map, and it is now the work list.** Reinforcement quantities
+  (`rebarBbs`, `rebarBbsCsvUrl` — deliberately split from the ACI check, which is a different
+  question); development budget, draws and billing (11); portfolio (3); saved views (3);
+  material palette (3); site and test fit (4); four unrelated singles. Five are marked as
+  **staying** — `enumOptions`, `searchAll`, `attachmentUrl`, `templates`, `rvtBridgeStatus` are
+  client-level, not domain.
 
   **⓽ took the AI-drafting banner apart** — and it is the clearest case yet for grouping by
   what methods ANSWER. One `// --- AI drafting ---` section held **three** questions: the
